+++ src/coverlet.collector/DataCollection/CoverageWrapper.cs
@@ -1,6 +1,6 @@
 ﻿using Coverlet.Collector.Utilities.Interfaces;
 using Coverlet.Core;
-using Coverlet.Core.Abstracts;
+using Coverlet.Core.Abstractions;
 using Coverlet.Core.Extensions;
 
 namespace Coverlet.Collector.DataCollection
