+++ src/coverlet.core/Helpers/Console.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-using Coverlet.Core.Abstracts;
+using Coverlet.Core.Abstractions;
 
 namespace Coverlet.Core.Helpers
 {
