+++ src/coverlet.core/Helpers/InstrumentationHelper.cs
@@ -9,7 +9,7 @@ using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
 using System.Text.RegularExpressions;
 
-using Coverlet.Core.Abstracts;
+using Coverlet.Core.Abstractions;
 
 namespace Coverlet.Core.Helpers
 {
