+++ test/coverlet.collector.tests/CoverletCoverageDataCollectorTests.cs
@@ -13,7 +13,7 @@ using Coverlet.Collector.Utilities;
 using Xunit;
 using Coverlet.Collector.DataCollection;
 using Coverlet.Core.Reporters;
-using Coverlet.Core.Abstracts;
+using Coverlet.Core.Abstractions;
 using Coverlet.Core.Helpers;
 using Microsoft.Extensions.DependencyInjection;
