+++ test/coverlet.core.tests/Reporters/Reporters.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-using Coverlet.Core.Abstracts;
+using Coverlet.Core.Abstractions;
 using Coverlet.MSbuild.Tasks;
 using Moq;
 using Xunit;
